Align encoder DRED analysis int widths

+++ decoder_osce_lace_apply.go
@@ -133,7 +133,7 @@ func (d *Decoder) applyOSCELACEMonoChannelWithControl(native []int16, mode osceL
 			fc.GainsQ16[sf] = ctrl.GainsQ16[sf]
 			fc.PitchL[sf] = ctrl.PitchL[sf]
 		}
-		fc.SignalType = ctrl.SignalType
+		fc.SignalType = int32(ctrl.SignalType)
 		numBits := ctrl.NumBits
 		if numBits < 0 {
 			numBits = 0
